POST and DELETE method updated

+++ functions/HttpTriggerUserIdTaskId/__init__.py
@@ -6,6 +6,7 @@ import azure.functions as func
 from ..Utils.dbHandler import dbHandler
 from ..Utils.ExceptionWithStatusCode import ExceptionWithStatusCode
 
+
 def main(req: func.HttpRequest) -> func.HttpResponse:
     logging.info(
         'Python HTTP trigger for /users/:userId/tasks/:taskId function is processing a request.')
@@ -45,6 +46,9 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
             getTaskById = getTask(cursor, row)
             logging.debug("Task retrieved successfully!")
             return getTaskById
+        elif method == "POST":
+            logging.debug("Attempting to create task...")
+            logging.debug("This method is not allowed!")
         elif method == "PUT":
             logging.debug("Attempting to update taskId: " + taskId)
             task_req_body = req.get_json()
@@ -58,12 +62,13 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
             logging.warn(f"Request with method {method} has been recieved, but that is not allowed for this endpoint")
             return func.HttpResponse(status_code=405)
 
-    #displays errors encountered when API methods were called
+    # displays errors encountered when API methods were called
     except ExceptionWithStatusCode as err:
         return func.HttpResponse(str(err), status_code=err.status_code)
     finally:
         conn.close()
         logging.debug('Connection to DB closed')
+
 
 def getTask(cursor, row):
     logging.debug('Retrieving task by Id...')
@@ -77,6 +82,17 @@ def getTask(cursor, row):
         mimetype='application/json'
     )
 
+# For POST and PATCH
+
+
+def methodNotAllowed():
+    logging.debug("This method is not Implemented")
+    return func.HttpResponse(
+        "Method not allowed!",
+        status_code=405
+    )
+
+
 def updateTask(cursor, task_req_body):
     task_req_body = req.get_json()
 
@@ -85,7 +101,8 @@ def updateTask(cursor, task_req_body):
         assert 'title' in task_req_body, 'User did not provide required field: title'
         assert 'taskDescription' in task_req_body, 'User did not provide required field: taskDescription'
     except AssertionError as err:
-        logging.error('User request body did not contain the necessary fields!')
+        logging.error(
+            'User request body did not contain the necessary fields!')
         return func.HttpResponse(
             'Task updated',
             status_code=200
@@ -93,18 +110,28 @@ def updateTask(cursor, task_req_body):
 
 
 def deleteTask(cursor, taskId):
-    logging.debug('Attempting to delete taskId: ' + taskId)
-    delete_task_query = 'DELETE FROM tasks WHERE taskId={}'.format(taskId)
-    logging.debug('Executing delete query')
-    cursor.execute(delete_task_query, (taskId))
-    logging.debug('task deleted')
-    return func.HttpResponse(
-        'task deleted',
-        status_code=200
-    )
+    row = get_task_row(cursor, taskId)
+    if not row:
+        logging.debug('Task not found')
+        return func.HttpResponse(
+            'Task not found',
+            status_code=404
+        )
+    else:
+        logging.debug('Attempting to delete taskId: ' + taskId)
+        delete_task_query = 'DELETE FROM tasks WHERE taskId={}'.format(taskId)
+        logging.debug('Executing delete query')
+        cursor.execute(delete_task_query, (taskId))
+        logging.debug('Task was successfully deleted')
+        return func.HttpResponse(
+            'Task deleted',
+            status_code=200
+        )
+
 
 def get_task_row(cursor, taskId):
     cursor.execute(
-        'SELECT taskId, taskUserId, title, taskDescription FROM tasks WHERE taskId={}'.format(taskId)
+        'SELECT taskId, taskUserId, title, taskDescription FROM tasks WHERE taskId={}'.format(
+            taskId)
     )
     return cursor.fetchone()
